Updated all the environment to work with Airsim & Tensorflow.
collision_testing.py: Decreased the acceleration to visualize the break.

With the original value of 0.65 the car still collide even after
pressing the break because of its hight velocity so I changed it to 0.1
and now the car actually stops before the collision.

SET_ENV.md: Wrote a "how to" file to set the environment.

One of the steps is to pip install all packages inside requirements.txt
that was added to this commit too.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>