Training seems to work corectly.
driving_training.py: Training the network.

For now the network have a single hidden layer of 10 nodes, and give an
accuracy of 0.02 on the dataset, we still can't check how the network
drive so at this stage we don't know if it is that bad.

random_collection.py: Create the dataset for the training.

data_dir/: Contains all the data files created by driving_training.py

Signed-off-by: Yoni Bettan <ybettan@redhat.com>